rm preprocess_constraints step
